+++ ext/FlexiChainsDynamicPPLExt.jl
@@ -139,7 +139,8 @@ end
         params_and_stats::AbstractMatrix{<:DynamicPPL.VarNamedTuple}
     )::VNChain
 
-Convert a matrix of [`DynamicPPL.VarNamedTuple`](@extref) to a `VNChain`.
+Convert a matrix of [`DynamicPPL.VarNamedTuple`](@extref
+`DynamicPPL.VarNamedTuples.VarNamedTuple) to a `VNChain`.
 """
 function AbstractMCMC.from_samples(
         ::Type{<:VNChain}, vnts::AbstractMatrix{<:DynamicPPL.VarNamedTuple}