+++ HISTORY.md
@@ -1,6 +1,6 @@
 # 0.5.1
 
-Added the `divergences` and `divergences_kwargs` keyword argument to `pairplot` to additionally superimpose a scatter plot of divergent transitions.
+Added the `divergences` and `divergences_kwargs` keyword arguments to `pairplot` to additionally superimpose a scatter plot of divergent transitions, as well as a keyword argument `args` that allows passing extra positional arguments to `pairplot`.
 
 # 0.5.0